requirements.txt
Django>=5.1
Pillow>=11.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
gunicorn>=23.0.0

django-allauth>=65.0.0
django-ckeditor-5>=0.2.18
django-threadedcomments>=2.0
django-notifications-hq>=1.8.3
django-push-notifications>=3.1.0
django-pwa>=2.0.0
django-meta>=2.4.0
django-recaptcha>=4.0.0
django-dbbackup>=4.2.0
django-taggit>=6.1.0

google-generativeai>=0.8.3
tweepy>=4.14.0
boto3>=1.35.0

django-htmx>=1.19.0
Django
django-allauth
django-htmx
django-pwa
django-taggit
gunicorn
Pillow
psycopg2-binary
python-dotenv
boto3
django-ckeditor-5
django-dbbackup
django-meta
django-notifications-hq
django-push-notifications
django-recaptcha
django-threadedcomments
google-generativeai
tweepy
PyJWT
cryptography
boto3>=1.35.0
cryptography
Django>=5.1
django-allauth>=65.0.0
django-ckeditor-5>=0.2.18
django-dbbackup>=4.2.0
django-htmx>=1.19.0
django-meta>=2.4.0
django-notifications-hq>=1.8.3
django-push-notifications>=3.1.0
django-pwa>=2.0.0
django-recaptcha>=4.0.0
django-taggit>=6.1.0
django-threadedcomments>=2.0
google-generativeai>=0.8.3
gunicorn>=23.0.0
Pillow>=11.0
psycopg2-binary>=2.9.9
PyJWT
python-dotenv>=1.0.0
tweepy>=4.14.0
boto3>=1.35.0
cryptography
Django>=5.1
django-allauth>=65.0.0
django-ckeditor-5>=0.2.18
django-dbbackup>=4.2.0
django-htmx>=1.19.0
django-meta>=2.4.0
django-notifications-hq>=1.8.3
django-push-notifications>=3.1.0
django-pwa>=2.0.0
django-recaptcha>=4.0.0
django-taggit>=6.1.0
django-threadedcomments>=2.0
google-generativeai>=0.8.3
gunicorn>=23.0.0
Pillow>=11.0
psycopg2-binary>=2.9.9
PyJWT
python-dotenv>=1.0.0
tweepy>=4.14.0
django-meta
